Fix build of pecl-intl
Fix the build of pecl-intl port on FreeBSD 12.1-RELEASE-p9 as of 2021-01-03.
Probably the compiler/port system has changed in a way, that TRUE and
FALSE defines are not available for pecl-intl. I have added two patches
to define TRUE and FALSE for all C files (in config.h) and one C++ file.

Fixes the following issues:

/opt/ports/php56/devel/pecl-intl/work-php56/intl-3.0.0/collator/collator_sort.c:346:26: error: use of undeclared identifier 'TRUE'
        collator_sort_internal( TRUE, INTERNAL_FUNCTION_PARAM_PASSTHRU );

/opt/ports/php56/devel/pecl-intl/work-php56/intl-3.0.0/breakiterator/codepointiterator_internal.cpp:58:53: error: use of undeclared identifier 'FALSE'
        this->fText = utext_clone(this->fText, that.fText, FALSE, TRUE, &uec);